303: Adapt cart to Storefront UI v0.7.x
