+++ components/organisms/o-microcart.vue
@@ -1,252 +1,52 @@
 <template>
-  <transition name="fade" mode="out-in">
-    <div v-if="totalItems" key="my-cart" class="o-microcart">
-      <h3 class="o-microcart__total-items">
-        {{ $t("Total items") }}: {{ productsCount }}
-      </h3>
-      <div class="collected-product-list">
-        <transition-group name="fade" tag="div">
-          <SfCollectedProduct
-            v-for="product in productsInCart"
-            :key="product.id"
-            :image="getThumbnailForProductExtend(product)"
-            :title="product.name"
-            :regular-price="getProductPrice(product).regular"
-            :special-price="getProductPrice(product).special"
-            :stock="10"
-            :qty="product.qty"
-            class="collected-product"
-            @click:remove="removeHandler(product)"
-            @input="changeQuantity(product, $event)"
-          >
-            <template #configuration>
-              <div class="collected-product__properties">
-                <SfProperty :name="$t('SKU')" :value="product.sku" />
-                <SfProperty
-                  v-for="property in getProductOptions(product)"
-                  :key="property.label"
-                  :name="property.label"
-                >
-                  <template #value>
-                    <span class="sf-property__value" v-html="property.value" />
-                  </template>
-                </SfProperty>
-              </div>
-            </template>
-          </SfCollectedProduct>
-        </transition-group>
-      </div>
-      <SfProperty
-        class="sf-property--full-width o-microcart__total-price"
-        v-for="total in totals"
-        :key="total.code"
-      >
-        <template #name>
-          <span class="sf-property__name">{{ total.title }}</span>
-        </template>
-        <template #value>
-          <SfPrice :regular="total.value | price" class="sf-price--big" />
-        </template>
-      </SfProperty>
-      <SfButton class="sf-button--full-width cart-action" @click.native="goToCheckout">
-        {{ $t("Go to checkout") }}
-      </SfButton>
-    </div>
-    <div v-else key="empty-cart" class="empty-cart">
-      <div class="empty-cart__banner">
-        <img
-          src="@storefront-ui/shared/icons/empty_cart.svg"
-          alt=""
-          class="empty-cart__icon"
-        >
-        <h3 class="empty-cart__label">
-          {{ $t("Your bag is empty") }}
-        </h3>
-        <p class="empty-cart__description">
-          {{
-            $t(
-              "Looks like you haven’t added any items to the bag yet. Start shopping to fill it in."
-            )
-          }}
-        </p>
-      </div>
-      <SfButton
-        class="sf-button--full-width color-secondary cart-action"
-        @click.native="startShopping"
-      >
-        {{ $t("Start shopping") }}
-      </SfButton>
-    </div>
-  </transition>
+  <component
+    v-if="isMicrocartOpen || isMicrocartComponentLoaded"
+    :is="theMicrocartAsyncComponent"
+    @close="$store.commit('ui/setMicrocart')"
+    @reload="reloadComponent()"
+  />
 </template>
 
 <script>
-import { mapGetters } from 'vuex';
-import { localizedRoute } from '@vue-storefront/core/lib/multistore';
-import { onlineHelper } from '@vue-storefront/core/helpers';
-import { getThumbnailForProduct } from '@vue-storefront/core/modules/cart/helpers';
-import { getProductPrice, getProductPriceFromTotals } from 'theme/helpers';
-import VueOfflineMixin from 'vue-offline/mixin';
-import onEscapePress from '@vue-storefront/core/mixins/onEscapePress';
+import { mapState } from 'vuex';
+import ALoadingSpinner from 'theme/components/atoms/a-loading-spinner';
+import ALoadingError from 'theme/components/atoms/a-loading-error';
 
-import {
-  SfButton,
-  SfCollectedProduct,
-  SfProperty,
-  SfPrice
-} from '@storefront-ui/vue';
+const OMicrocartPanel = () =>
+  import(/* webpackChunkName: "vsf-microcart" */ 'theme/components/organisms/o-microcart-panel');
 
 export default {
-  components: {
-    SfButton,
-    SfCollectedProduct,
-    SfProperty,
-    SfPrice
+  data () {
+    return {
+      isMicrocartComponentLoaded: false,
+      theMicrocartAsyncComponent: () => ({
+        component: OMicrocartPanel().then(component => {
+          this.isMicrocartComponentLoaded = true;
+          return component;
+        }),
+        loading: ALoadingSpinner,
+        error: ALoadingError,
+        timeout: 3000
+      })
+    };
   },
-  mixins: [VueOfflineMixin, onEscapePress],
   computed: {
-    ...mapGetters({
-      productsInCart: 'cart/getCartItems',
-      totals: 'cart/getTotals'
-    }),
-    totalItems () {
-      return this.productsInCart.length;
-    },
-    productsCount () {
-      let count = 0;
-      this.productsInCart.forEach(product => {
-        count = count + parseInt(product.qty);
-      });
-      return count;
-    }
+    ...mapState({
+      isMicrocartOpen: state => state.ui.microcart
+    })
   },
   methods: {
-    closeMicrocartExtend () {
-      this.$store.dispatch('ui/toggleMicrocart');
-    },
-    onEscapePress () {
-      this.closeMicrocartExtend();
-    },
-    startShopping () {
-      this.$router.push(localizedRoute('/'));
-      this.closeMicrocartExtend();
-    },
-    getThumbnailForProductExtend (product) {
-      return getThumbnailForProduct(product);
-    },
-    getProductPrice (product) {
-      return onlineHelper.isOnline && product.totals && product.totals.options
-        ? getProductPriceFromTotals(product)
-        : getProductPrice(product);
-    },
-    getProductOptions (product) {
-      return onlineHelper.isOnline && product.totals && product.totals.options
-        ? product.totals.options
-        : product.options || {};
-    },
-    removeHandler (product) {
-      this.$store.dispatch('cart/removeItem', { product: product });
-    },
-    goToCheckout () {
-      this.$router.push(localizedRoute('/checkout'));
-      this.closeMicrocartExtend();
-    },
-    changeQuantity (product, newQuantity) {
-      this.$store.dispatch('cart/updateQuantity', {
-        product: product,
-        qty: newQuantity
+    reloadComponent () {
+      this.theMicrocartAsyncComponent = () => ({
+        component: OMicrocartPanel().then(component => {
+          this.isMicrocartComponentLoaded = true;
+          return component;
+        }),
+        loading: ALoadingSpinner,
+        error: ALoadingError,
+        timeout: 3000
       });
-    },
-    onQuantityChange () {
-      // Unfortunately $forceUpdate() is needed here because `totals` key in cart items
-      // is added but not in a reactive way, so Vue is not able to detect this change and
-      // re-render our view.
-      // The callback for 'cart-after-itemchanged' event can be removed when the following PR
-      // will be merged in VSF: https://github.com/DivanteLtd/vue-storefront/pull/4079/
-      this.$forceUpdate();
     }
-  },
-  beforeMount () {
-    this.$bus.$on('cart-after-itemchanged', this.onQuantityChange);
-  },
-  beforeDestroy () {
-    this.$bus.$off('cart-after-itemchanged', this.onQuantityChange);
   }
 };
 </script>
-
-<style lang="scss" scoped>
-@import "~@storefront-ui/shared/styles/helpers/breakpoints";
-
-#cart {
-  box-sizing: border-box;
-  @include for-desktop {
-    max-width: 1272px;
-    margin: auto;
-  }
-}
-.o-microcart {
-  flex: 1;
-  display: flex;
-  flex-direction: column;
-  &__total-items {
-    font-family: var(--font-family-secondary);
-    font-size: var(--font-lg);
-    font-weight: var(--font-light);
-  }
-  &__total-price {
-    margin-bottom: calc(var(--spacer-xl) / 2);
-  }
-}
-.collected-product-list {
-  flex: 1;
-  margin: var(--spacer-xl) var(--spacer-xl);
-}
-.collected-product {
-  margin: var(--spacer-xl) 0;
-  --collected-product-max-width: none;
-  --collected-product-image-background: var(--c-white);
-  &__properties {
-    margin-top: var(--spacer-xl);
-    --property-name-font-size: 0.8rem;
-    --property-value-font-font-size: 0.8rem;
-  }
-}
-.empty-cart {
-  flex: 1;
-  display: flex;
-  flex-direction: column;
-  &__banner {
-    flex: 1;
-    display: flex;
-    flex-direction: column;
-    justify-content: center;
-    align-items: center;
-  }
-  &__icon {
-    width: 18.125rem;
-    height: 12.3125rem;
-    margin-left: 60%;
-    @include for-desktop {
-      margin-left: 50%;
-    }
-  }
-  &__label,
-  &__description {
-    line-height: 1.6;
-    text-align: center;
-  }
-  &__label {
-    margin-top: var(--spacer-2xl);
-    font-size: var(--font-lg);
-  }
-  &__description {
-    margin-top: var(--spacer-xl);
-  }
-}
-@include for-mobile {
-  .cart-action {
-    margin-bottom: var(--spacer-xl);
-  }
-}
-</style>
